Hi all,

This pull request contains a backport of commit [47df1459](https://github.com/openjdk/jdk/commit/47df14590c003ccb1607ec0edfe999fcf2aebd86) from the [openjdk/jdk](https://git.openjdk.org/jdk) repository.

The commit being backported was authored by Amit Kumar on 10 Apr 2024 and was reviewed by Lutz Schmidt and Martin Doerr.

Thanks!


Tested `{tier1} X {fastdebug}` with `-XX:-TieredCompilation` and do not see any new failure appearing. 

Benchmarking Result: 

```
./build/linux-s390x-server-release/jdk/bin/java -Xms4g -Xmx4g -jar dacapo-9.12-MR1-bach.jar h2 -s huge -t 1 -n 1

With Patch: 
===== DaCapo 9.12-MR1 h2 PASSED in 148008 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 152460 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 151682 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 146845 msec =====


Without Patch: 
===== DaCapo 9.12-MR1 h2 PASSED in 192967 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 206936 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 191804 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 196735 msec =====

```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] [JDK-8310513](https://bugs.openjdk.org/browse/JDK-8310513) needs maintainer approval

### Issue
 * [JDK-8310513](https://bugs.openjdk.org/browse/JDK-8310513): [s390x] Intrinsify recursive ObjectMonitor locking (**Enhancement** - P4 - Approved)


### Reviewers
 * [Lutz Schmidt](https://openjdk.org/census#lucy) (@RealLucy - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk22u.git pull/134/head:pull/134` \
`$ git checkout pull/134`

Update a local copy of the PR: \
`$ git checkout pull/134` \
`$ git pull https://git.openjdk.org/jdk22u.git pull/134/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 134`

View PR using the GUI difftool: \
`$ git pr show -t 134`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk22u/pull/134.diff">https://git.openjdk.org/jdk22u/pull/134.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk22u/pull/134#issuecomment-2046507197)