@offamitkumar 2024-04-29
/approval request Need to backport this change to JDK-22 as other archs ported it in JDK-21 and for them it's available in JDK-22u already.